requirements.txt
flask
flask-cors
python-dotenv
SQLAlchemy
certifi
langchain
langchain-core
bcrypt
PyJWT
waitress
crewai
litellm
json-repair
